requirements.txt
# Federated Learning Framework
flwr==1.7.0
flwr-datasets==0.0.2
protobuf==4.25.8

# Deep Learning
tensorflow==2.16.1
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2

# Backend API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Authentication
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
aiosqlite==0.19.0

# Communication
websockets==12.0
requests==2.32.5
grpcio==1.60.0
grpcio-tools==1.60.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1


# # Federated Learning Framework
# flwr==1.7.0
# flwr-datasets==0.0.2

# # Deep Learning
# tensorflow==2.16.1
# numpy==1.26.4
# pandas==2.0.3
# scikit-learn==1.5.2

# # Backend API
# fastapi==0.109.0
# uvicorn[standard]==0.27.0
# pydantic==2.5.3
# python-multipart==0.0.6

# # Authentication
# PyJWT==2.8.0
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4

# # Database
# aiosqlite==0.19.0

# # Communication
# websockets==12.0
# requests==2.31.0
# grpcio==1.60.0
# grpcio-tools==1.60.0

# # Utilities
# python-dotenv==1.0.0
# pyyaml==6.0.1